% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_along.R
\name{place_along}
\alias{place_along}
\title{Place along}
\usage{
place_along(label, x, y, offset, dist = 0.5, vpos = 0.2, win = 1,
  centred = T)
}
\arguments{
\item{centred}{}
}
\description{
Place along
}
