% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_along.R
\name{point_along}
\alias{point_along}
\title{Point along}
\usage{
point_along(d, x, y)
}
\arguments{
\item{d}{linear distance of the researched point along the path from the first point.}

\item{x}{vector of the x-position of the points along the path.}

\item{y}{vector of the y-position of the points along the path.}
}
\value{
vector of (x,y) coordinates of the reasearch point. Returns extremities if the distance in negative or null, or too long.
}
\description{
Find a point along a path
}
\examples{

x <- c(1,1,1,2,2)
y <- 1:5

d <- 1
point_along(d, x, y)
d <- 4.123
point_along(d, x, y)
d <- 5
point_along(d, x, y)
}
