% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_on.R
\name{point_on}
\alias{point_on}
\title{Point on}
\usage{
point_on(d, x, y)
}
\arguments{
\item{d}{distance to the first point.}

\item{x}{vector of the x-position of the points along the path.}

\item{y}{vector of the y-position of the points along the path.}
}
\description{
Find the coordinates of a point to a flight linear distance from the first point of the path.
}
\examples{

x <- sort(floor(runif(10)*100)/20)
y <- sort(floor(runif(10)*100)/20)
d <- mean(x)
point <- point_on(d, x, y)
ggplot(data.frame(x,y), aes(x,y)) + geom_path() + geom_point(data= data.frame(x = point[1], y = point[2]))
}
