% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_letter.R
\name{position_letter}
\alias{position_letter}
\title{Position letter}
\usage{
position_letter(x, y, xr, yr, win = 1, offset = 0.1, centred = T)
}
\arguments{
\item{centred}{}
}
\description{
Position letter
}
