% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_river_label.R
\name{geom_river_label}
\alias{geom_river_label}
\title{Text}
\usage{
geom_river_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., parse = FALSE, nudge_x = 0, nudge_y = 0,
  na.rm = FALSE, centred = TRUE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{parse}{If `TRUE`, the labels will be parsed into expressions and
displayed as described in `?plotmath`.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels by.
Useful for offsetting text from points, particularly on discrete scales.}
}
\description{
Text
}
\section{`geom_label`}{

Currently `geom_label` does not support the `angle` aesthetic and
is considerably slower than `geom_text`. The `fill` aesthetic
controls the background colour of the label.
}

\section{Alignment}{

You can modify text alignment with the `vjust` and `hjust`
aesthetics. These can either be a number between 0 (right/bottom) and
1 (top/left) or a character (`"left"`, `"middle"`, `"right"`, `"bottom"`,
`"center"`, `"top"`). There are two special alignments: `"inward"` and
`"outward"`. Inward always aligns text towards the center, and outward
aligns it away from the center.
}

\examples{

}
